+++ src/lib/modules/SATModule/Vec.h
@@ -207,7 +207,7 @@ namespace Minisat
         if( cap >= min_cap )
             return;
         int add = imax( (min_cap - cap + 1) & ~1, ((cap >> 1) + 2) & ~1 );    // NOTE: grow by approximately 3/2
-        if( (add > INT_MAX - cap || ((data = (T*)::realloc( data, (cap += add) * sizeof(T) )) == NULL)) && errno == ENOMEM )
+        if( (add > INT_MAX - cap || ((data = (T*)::realloc( data, (unsigned)(cap += add) * sizeof(T) )) == NULL)) && errno == ENOMEM )
             throw OutOfMemoryException();
     }
